Conflicts:
	Unity Project/Assets/Scenes/KC Scene.unity
	Unity Project/Assets/Scripts/Levels/GridSpace.cs
	Unity Project/Assets/Scripts/MANAGERS/LevelManager.cs
	Unity Project/Assets/Scripts/MANAGERS/StartManager.cs
	Unity Project/Assets/Scripts/WorldObjects/NPC/NPC.cs
	Unity Project/Assets/Scripts/WorldObjects/NPC/NPCInstance.cs

Former-commit-id: 53d1738ebc48539e887a92ab119e05af1a265e75

+++ Unity Project/Assets/Scripts/Library/Interfaces/PassesTurns.cs
@@ -1,7 +1,7 @@
 public interface PassesTurns
 {
     void UpdateTurn();
-    int CurrentPoints { get; set; }
-    int BasePoints { get; set; }
+    ulong CurrentPoints { get; set; }
+    ulong BasePoints { get; set; }
     bool IsActive { get; set; }
 }
